+++ ricaun.DA4R.NUnit.Tests/ProjectInfoTests.cs
@@ -8,9 +8,11 @@ namespace ricaun.DA4R.NUnit.Tests
         interface IElement<T> where T : Element { }
     }
 
+#if REVIT2021_OR_GREATER
     public class ForgeTypeIdTests
     {
         interface IForgeTypeId : IElement<ForgeTypeId> { }
         interface IElement<T> where T : ForgeTypeId { }
     }
+#endif
 }