+++ ricaun.DA4R.NUnit/RevitExtensions/ControlledApplicationExtension.cs
@@ -17,10 +17,10 @@
         {
             var type = typeof(ControlledApplication);
 
-            var propertie = type.GetFields(BindingFlags.Instance | BindingFlags.NonPublic)
+            var property = type.GetFields(BindingFlags.Instance | BindingFlags.NonPublic)
                 .FirstOrDefault(e => e.FieldType == typeof(Application));
 
-            return propertie?.GetValue(application) as Application;
+            return property?.GetValue(application) as Application;
         }
 
         /// <summary>